+++ controll-app/src/devices/pc/KSQPCController.cpp
@@ -153,7 +153,7 @@ KSQPCController::columnCount(const QModelIndex &parent) const {
 //-----------------------------------------------------------------------------
 QVariant
 KSQPCController::data(const QModelIndex &index, int role) const {
-    if (index.row() < m_pcs.size()) {
+    if (index.row() < static_cast<int>(m_pcs.size())) {
         auto l = *std::next(m_pcs.begin(), index.row());
 
         switch (role) {
